refactor: TransactionService handles transactions not AccountService
- The deposit and withdraw transactions are not handle by
  AccountService anymore. They are now in TransactionService.
  AccountController now uses TransactionService in addition to
  AccountService. The lines where AccoutService was used for
  the deposit and withdraw requests now use TransactionService.
