+++ backend/src/main/java/dev/jordanjoseph/backend/controller/AccountController.java
@@ -3,6 +3,7 @@ package dev.jordanjoseph.backend.controller;
 import dev.jordanjoseph.backend.dto.AccountView;
 import dev.jordanjoseph.backend.dto.MoneyRequest;
 import dev.jordanjoseph.backend.service.AccountService;
+import dev.jordanjoseph.backend.service.TransactionService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,6 +17,9 @@ public class AccountController {
     @Autowired
     private AccountService accountService;
 
+    @Autowired
+    private TransactionService transactionService;
+
     @GetMapping("/me")
     public AccountView me() {
         return accountService.myAccount();
@@ -23,12 +27,12 @@ public class AccountController {
 
     @PostMapping("/{id}/deposit")
     public AccountView deposit(@PathVariable UUID id, @RequestBody @Valid MoneyRequest request) {
-        return accountService.deposit(id, request.amount());
+        return transactionService.deposit(id, request.amount());
     }
 
     @PostMapping("/{id}/withdraw")
     public AccountView withdraw(@PathVariable UUID id, @RequestBody @Valid MoneyRequest request) {
-        return accountService.withdraw(id, request.amount());
+        return transactionService.withdraw(id, request.amount());
     }
 
 }
